Delete targets

Delete targets. Deletes both file and object targets, and removes their entries from the remake database. Using make("clean") should probably be the general way to clean up, but this might be useful if you have specific objects to delete. While files can be deleted in this way, deleting in the file system is also fine.

delete(target_names, dependencies = FALSE, verbose = TRUE,
  remake_file = "remake.yml")

Arguments

target_names

Names of targets to delete

dependencies

Delete dependencies of the target too? Use with caution.

verbose

Be verbose when loading the remake file and when deleting targets.

remake_file

Name of the remakefile (by default remake.yml)

Details

This function ignores cleanup_level and will quite happily delete things that have been flagged as cleanup_level: purge - be careful using dependencies=TRUE as this will delete all dependencies. See list_dependencies to see what would be deleted.

It is an error to try to delete a fake target (i.e., a target with no rule but that exists to group other dependencies). It is not an error to delete the dependencies of such a target.

If run with verbose=TRUE delete will print information about targets that are deleted with a DEL for each deleted target and an empty string if the target is already nonexistant.