Vector of file extensions

Returns the vector of default file extensions. If a target ends in one of these, then it will be considered a file, rather than an object.

file_extensions()

Details

To include additional file extensions, include them in the yaml like (at the top level):

file_extensions: ["phy", "tre"]

Any number of extensions can be listed. Don't use a leading period (it will be dropped). The union of file extensions listed here and in file_extensions() will be used, so if you accidently include a default extension (or if one is included in a future remake version) it is no problem.