Control remake verbosity

Helper function to set options for verbosity.

remake_verbose(verbose = getOption("remake.verbose", TRUE),
  noop = getOption("remake.verbose.noop", TRUE),
  command = getOption("remake.verbose.command", TRUE),
  command_abbreviate = getOption("remake.verbose.command.abbreviate", TRUE),
  target = NULL)

Arguments

verbose

Print progress at each step that remake does something.

noop

Print progress for steps that are non-operations, such as targets that need nothing done to them. Setting this to FALSE is useful for very large projects.

command

Print the command along with the progress information? This is only printed when remake actually runs something.

command_abbreviate

Abbreviate the command information so that it fits on one line. If FALSE then the command will be allowed to run on for as many lines as required.

target

Print information that the target produces (via message, cat or print). If FALSE then these messages will be suppressed.

Details

The first four options have a natural nesting: setting progress=FALSE prevents printing any progress information, so the value of noop, command and command_abbreviate does not matter. Similarly, setting command=FALSE means that command_abbreviate does not matter.