Automatically generate .gitignore

Attempts to add targets that remake will generate to your .gitignore. If the file already exists, then the files will be added (if not already present), otherwise a file will be created. If check_git is TRUE (the default) then we attempt to check with git to see if the files are already ignored by any gitignore (including a system-specific global gitignore) and only files that are not already excluded will be added. Probably after running this function you will want to do some editing.

auto_gitignore(remake_file = "remake.yml", check_git = TRUE,
  dry_run = FALSE)

Arguments

remake_file

Name of the remake file to use, by default remake.yml.

check_git

Use the output of git check-ignore to determine which files are already ignored? This is only done if git is detected to be installed and if the working directory appears to be in a git repository.

dry_run

Don't modify the .gitignore, but instead return a character vector of what would be added.