Dump remake contents to environment

Dump the contents of remake into an environment; by default the global environment. This is similar in effect to create_bindings but does not create links; instead a copy of everything that remake has built, plus all functions sources into remake, are copied into the environment.

dump_environment(envir = .GlobalEnv, verbose = TRUE,
  allow_missing_packages = FALSE, remake_file = "remake.yml")

Arguments

envir

Environment to copy into; by default the global environment.

verbose

Be verbose when loading the remakefile

allow_missing_packages

Allow missing packages when loading remake file?

remake_file

Name of the remakefile (by default remake.yml)