Fetch last computed

Fetch the last computed value from the remake database.

fetch(target_name, require_current = FALSE, verbose = TRUE,
  allow_missing_packages = FALSE, remake_file = "remake.yml")

Arguments

target_name

The name of a single target to fetch the value of

require_current

Logical indicating if the targets must be up-to-date to be fetched. If this is TRUE and the targets are not up-to-date, then an error will be thrown.

verbose

Be verbose when loading remake file? Default is TRUE.

allow_missing_packages

Allow missing packages when loading remake file?

remake_file

Name of the remakefile (by default remake.yml)

Value

An R object.

Details

The last computed value would be returned invisibly by make, but this function provides a way of accessing values without ever triggering a rebuild. As such, it's possible that the target is not made, or is not current, so there are options for controlling what to do in this case.

It is an error to use this function with file targets (but see is_current for checking currentness) and fetch_archive for extracting files from archives.