Install missing packages

Install missing packages

install_missing_packages(remake_file = "remake.yml", instructions = FALSE,
  missing_only = TRUE, skip_target_packages = FALSE)

Arguments

remake_file

Name of the remakefile to look for the list of required packages

instructions

Rather than install anything, prints instructions on how to install missing things

missing_only

If FALSE, install everything, rather than just missing packages. This might be useful to set to TRUE in conjunction with instructions=TRUE to generate a full list to install.

skip_target_packages

Skip packages that are mentioned only in targets?

Details

No version comparison is done - see packrat for a more complete package management solution, though in my initial experiments it does not play that nicely with remake.

Author
Rich FitzJohn