Install running script.

Install running script to a local directory. This directory should be on the $PATH. Once this has been done, you can run remake with `remake`. See `remake --help` for more information, or a summary below.

install_remake(destination_directory, overwrite = FALSE)

Arguments

destination_directory

Directory to install `remake` to. Should be on your path, though the current directory may be useful too. The file will be installed as file.path(dest, "remake")

overwrite

Overwrite an existing file?

Details

The installed script is just a wrapper to an internal remake function, designed so that the wrapper script does not need to be installed after upgrading remake.

Because of the design of Rscript, the helper script loads the method package: in my experience many things just do not work without that package loaded and Rscript does not load it by default (especially in conjunction with ::).

This remake script accepts standard unix-style command-line options to drive a subset of remake's functionality. Supported featues are:

Running remake::make, as remake target1 [target2]. If targets are ommitted the default target will be used.

listing targets, as --list-targets, which calls list_targets

generating a script with -s or --script (printing to standard output) or --script-file prints to a file.

return the version, as -v or --version, returning packageVersion("remake")

Additionally, the file used can be selected by using -f or --file (following make's convention), and remake can be run with verbose=FALSE by passing in -q or --quiet.