List contents of a remake archive

List contents of a remake archive

list_archive(archive_file = "remake.zip", detail = FALSE)

Arguments

archive_file

Name of the zip file to read from, by default "remake.zip".

detail

Return a data frame with more detail?

Value

A character vector with the contents of the archive.

Details

NOTE: At present this unzips the entire archive.