Command line interface to remake

This function exists to support commandline use, as installed by install_remake. This little script accepts standard unix-style command-line options to drive a subset of remake's functionality. Don't run this function from within R! This uses commandArgs to pass along arguments to various remake functions.

Details

See remake --help for instructions.

At the moment, supported featues are:

Running remake::make, as remake target1 [target2]. If targets are ommitted the default target will be used.

listing targets, as --list-targets, which calls list_targets

generating a script with -s or --script (printing to standard output) or --script-file prints to a file.

return the version, as -v or --version, returning packageVersion("remake")

Additionally, the file used can be selected by using -f or --file (following make's convention), and remake can be run with verbose=FALSE by passing in -q or --quiet.